+++ components/Project/index.js
@@ -1,0 +1,4 @@
+export { default as Project } from './Project'
+export { default as DownloadBlock } from './DownloadBlock'
+export { default as BookCreate } from './BookCreate'
+export { default as ChapterList } from './ChapterList'
